Prepare decimals count update, and new single token

+++ blockchain/contracts/badges/cost/IFractionCostBadges.sol
@@ -10,10 +10,10 @@ interface IFractionCostBadges is IPausable {
     /**
      * @dev Update the cost badge of the given f nft id
      */
-    function updateBadge(uint256 _fractionId, uint64 _coefficient) external;
+    function updateBadge(uint256 _fractionId, uint128 _coefficient) external;
 
     /**
      * @dev Find the badge for the given f nft id
      */
-    function getBadge(uint256 _fractionId) external view returns (uint64);
+    function getBadge(uint256 _fractionId) external view returns (uint128);
 }
